+++ examples/09_custom_steps/main.cpp
@@ -1,0 +1,6 @@
+#include "goodbye.hpp"
+
+int main() {
+  goodbye();
+  return 0;
+}